Add files via upload
This implementation of the percolation system uses generic programming in PercolationStats.hpp to flexibly decide which Union-Find variant to use. because of  this, I added default constructors to the percolation implementations and made PercolationType a private member of the stats class, initializing it within the constructor. This way, a fresh percolation object is created for each trial, keeping the statistics logic decoupled from the specific Union-Find choice. In hindsight, this is something I should have already been doing, since it allows Quick Find, Weighted Quick Union, or any other variant to be swapped in without changing the PercolationStats code.